+++ utilpaisagem/scenery/image_service.py
@@ -1,13 +1,17 @@
+from typing import MutableMapping
 from pathlib import Path
 from urllib import request
 from numbers import Number
+from collections import OrderedDict
 from urllib.error import URLError, ContentTooShortError
 from utilpaisagem.scenery.common import Coordinates, DOWNLOAD_RES, MIN_RES
 
 class ImageService(object):
     """
     Base class for image services. It has a `download` method, a `description` string,
-    a license_link string and an `availability_area` string.
+    a license_link string and an `availability_area` string. ImageService objects may
+    be compared and ordered by their names.
+
     Each child class should define its strings and a `_url` method, which returns a URL
     from `coordinates`, `width` and `height` parameters.
     """
@@ -17,6 +21,21 @@ class ImageService(object):
     license_link:str
     availability_area:str
     max_size:int = 2**DOWNLOAD_RES
+
+    def __gt__(self, other):
+        return self.name > other.name
+    
+    def __ge__(self, other):
+        return self.name >= other.name
+
+    def __eq__(self, other):
+        return self.name == other.name
+    
+    def __le__(self, other):
+        return self.name <= other.name
+    
+    def __lt__(self, other):
+        return self.name < other.name
 
     def _get_url(self, coordinates:Coordinates, width:int, height:int):
         """
@@ -87,9 +106,9 @@ class ImageService(object):
 class _ArcGIS(ImageService):
     def __init__(self):
         self.name = 'ArcGIS'
-        self.description = 'Worldwide service under restrictive license'
+        self.description = _('ArcGIS worldwide service under restrictive license')
         self.license_link = 'https://www.esri.com/en-us/legal/terms/full-master-agreement'
-        self.availability_area = 'Worldwide'
+        self.availability_area = _('worldwide')
         self.max_size = 4096
 
     def can_download(self, coordinates:Coordinates):
@@ -100,8 +119,62 @@ class _ArcGIS(ImageService):
     def _get_url(self, coordinates:Coordinates, width:int, height:int) -> str:
         return f'https://services.arcgisonline.com/arcgis/rest/services/World_Imagery/MapServer/export?bbox={coordinates.lon_left},{coordinates.lat_top},{coordinates.lon_right},{coordinates.lat_bottom}&bboxSR=4326&imageSR=4326&size={width},{height}&format=png24&f=image'
 
+class _PNOA(ImageService):
+    def __init__(self):
+        self.name = 'PNOA'
+        self.description = _('Plan Nacional de Ortofotografía Aérea by Instituto Geográfico Nacional (CC-BY 4.0)')
+        self.license_link = 'https://creativecommons.org/licenses/by/4.0/'
+        self.availability_area = _('Spain')
+    
+    def can_download(self, coordinates:Coordinates) -> bool:
+        # https://www.ign.es/wms-inspire/pnoa-ma?Request=GetCapabilities&Service=WMS
+        return coordinates.lon_left >= -19.0 and \
+            coordinates.lat_bottom >= 27.0 and \
+            coordinates.lon_right <= 5.0 and \
+            coordinates.lat_top <= 44.0
+    
+    def _get_url(self, coordinates:Coordinates, width:int, height:int) -> str:
+        # return f'https://www.ign.es/wms-inspire/pnoa-ma?SERVICE=WMS|VERSION=1.1.1|REQUEST=GetMap|LAYERS=OI.OrthoimageCoverage|SRS=EPSG:4326|BBOX={coordinates.lon_left},{coordinates.lat_top},{coordinates.lon_right},{coordinates.lat_bottom}|WIDTH={width}|HEIGHT={height}|FORMAT=image/png'
+        return f'https://www.ign.es/wms-inspire/pnoa-ma?SERVICE=WMS|VERSION=1.3.0|REQUEST=GetMap|LAYERS=OI.OrthoimageCoverage|CRS=EPSG:4326|BBOX={coordinates.lon_left},{coordinates.lat_top},{coordinates.lon_right},{coordinates.lat_bottom}|WIDTH={width}|HEIGHT={height}|FORMAT=image/png'
+
+class _USGS(ImageService):
+    def __init__(self):
+        self.name = 'USGS'
+        self.description = _('U.S. Geographical Surveys (public domain)')
+        self.license_link = 'https://www.usgs.gov/information-policies-and-instructions/copyrights-and-credits'
+        self.availability_area = _('USA')
+    
+    #TODO
+    def can_download(self, coordinates:Coordinates) -> bool:
+        return True
+    
+    def _get_url(self, coordinates:Coordinates, width:int, height:int) -> str:
+        return f'https://basemap.nationalmap.gov/arcgis/rest/services/USGSImageryOnly/MapServer/export?bbox={coordinates.lon_left},{coordinates.lat_top},{coordinates.lon_right},{coordinates.lat_bottom}&bboxSR=4326&size={height},{width}&imageSR=4326&format=png24&f=image'
+
+class _Bayern(ImageService):
+    def __init__(self):
+        self.name = 'Geobaisdaten Bayern'
+        self.description = _('Kostenfreie Geodaten der Bayerischen Vermessungsverwaltung')
+        self.license_link = 'https://creativecommons.org/licenses/by/4.0/deed.de'
+        self.availability_area = _('Bavaria (Deutschland)')
+
+    #TODO
+    def can_download(self, coordinates:Coordinates) -> bool:
+        return True
+    
+    def _get_url(self, coordinates:Coordinates, width:int, height:int) -> str:
+        return f''
+
 # There is no need for a singleton. This list is only a centralized place
 # for image service classes stored in order to facilitate GUI development.
-IMAGE_SERVICES = [
+_IMAGE_SERVICES = [
     _ArcGIS(),
+    # _PNOA(), # Not working. No photoscenery tool for FG can get its images
+    _USGS(),
+    # _Bayern(), # TODO
 ]
+
+IMAGE_SERVICES:MutableMapping[str, ImageService] = OrderedDict()
+_IMAGE_SERVICES.sort()
+for im in _IMAGE_SERVICES:
+    IMAGE_SERVICES[im.name] = im